feat: format token attributes
